front

+++ flask_base/src/services/ratings.py
@@ -3,7 +3,7 @@ import requests
 from sqlalchemy import exc
 from marshmallow import EXCLUDE
 from flask_login import current_user
-
+import services.songs as songs_service
 from schemas.rating import RatingSchema
 from models.http_exceptions import *
 
@@ -19,6 +19,12 @@ def create_rating(rating_register):
     rating_schema = RatingSchema().loads(json.dumps(rating_register), unknown=EXCLUDE)
     rating_schema["user_id"] = current_user.id
 
+    #check if music_id exists
+    songs, status_code = songs_service.get_song(rating_schema["music_id"])
+  
+
+    if status_code != 200:
+        return songs, status_code
     
     # on crée la chonson côté API ratings
     response = requests.request(method="POST", url=ratings_url, json=rating_schema)
@@ -39,21 +45,38 @@ def get_rating(id):
 
 def update_rating(id, rating_update):
     
-        # s'il y a quelque chose à changer côté API 
-        rating_schema = RatingSchema().loads(json.dumps(rating_update), unknown=EXCLUDE)
-        print(rating_schema)
-        response = None
-        if not RatingSchema.is_empty(rating_schema):
-            # on lance la requête de modification
-            response = requests.request(method="PUT", url=ratings_url+"/"+id, json=rating_schema)
-            print(response.status_code)
-            if response.status_code != 200:
-                return response.json(), response.status_code
     
-        return response.json(), response.status_code
+    # check if the rating is from the current user
+    rating, status_code = get_rating(id)
+    if status_code != 200:
+        return rating, status_code
     
+    if rating["user_id"] != current_user.id:
+        raise Forbidden
+    
+    # s'il y a quelque chose à changer côté API 
+    rating_schema = RatingSchema().loads(json.dumps(rating_update), unknown=EXCLUDE)
+    print(rating_schema)
+    response = None
+    if not RatingSchema.is_empty(rating_schema):
+        # on lance la requête de modification
+        response = requests.request(method="PUT", url=ratings_url+"/"+id, json=rating_schema)
+        print(response.status_code)
+        if response.status_code != 200:
+            return response.json(), response.status_code
+
+    return response.json(), response.status_code
+
 
 def delete_rating(id):
+    # check if the rating is from the current user
+    rating, status_code = get_rating(id)
+    if status_code != 200:
+        return rating, status_code
+    
+    if rating["user_id"] != current_user.id:
+        raise Forbidden
+    
     response = requests.request(method="DELETE", url=ratings_url+"/"+id)
     return response.json(), response.status_code
